Shapes: move double stroke highlight to a mixin
No sense in repeating code!
